+++ src/common/interfaces.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod';
+import { Logger } from '@map-colonies/js-logger';
 import type { ICreateTaskBody, IJobResponse, ITaskResponse } from '@map-colonies/mc-priority-queue';
 import {
   type InputFiles,
@@ -15,14 +16,14 @@ import {
 import type { BBox, Feature, FeatureCollection, MultiPolygon, Polygon } from 'geojson';
 import type { ITileRange } from '@map-colonies/mc-utils';
 import type { Span, SpanContext } from '@opentelemetry/api';
-import type { IngestionSwapUpdateFinalizeJob, IngestionUpdateFinalizeJob } from '../utils/zod/schemas/job.schema';
+import type { ExportFinalizeTask, ExportJob, IngestionSwapUpdateFinalizeJob, IngestionUpdateFinalizeJob } from '../utils/zod/schemas/job.schema';
 import {
   ingestionSwapUpdateFinalizeJobParamsSchema,
   ingestionUpdateFinalizeJobParamsSchema,
   extendedRasterLayerMetadataSchema,
   ingestionNewExtendedJobParamsSchema,
 } from '../utils/zod/schemas/jobParameters.schema';
-import { LayerCacheType, SeedMode, SqlDataType } from './constants';
+import { LayerCacheType, SeedMode } from './constants';
 
 export type StepKey<T> = keyof T & { [K in keyof T]: T[K] extends boolean ? K : never }[keyof T]; // this is a utility type that extracts the keys of T that are of type boolean
 
@@ -215,17 +216,28 @@ export interface ExportTaskParameters {
 
 export type ExportTask = ICreateTaskBody<ExportTaskParameters>;
 
+export type GpkgArtifactProperties = Omit<RasterLayerMetadata, 'productStatus' | 'footprint'>;
+export type JsonArtifactProperties = Omit<RasterLayerMetadata, 'productStatus'> & { sha256: string };
+
 //#endregion exportTask
 
-//#region gpkgClient
+//#region exportFinalizeTask
 
-export interface TableDefinition {
-  name: string;
-  dataType: SqlDataType;
-  value: unknown;
+export interface ExportFinalizeExecutionContext {
+  job: ExportJob;
+  task: ExportFinalizeTask;
+  paths: ExportFinalizeGpkgPaths;
+  telemetry: JobAndTaskTelemetry;
+  logger: Logger;
 }
 
-//#endregion gpkgClient
+export interface ExportFinalizeGpkgPaths {
+  gpkgFilePath: string;
+  gpkgRelativePath: string;
+  gpkgDirPath: string;
+}
+
+//#endregion exportFinalizeTask
 
 //#region mapproxyApi
 export interface PublishMapLayerRequest {