+++ src/job/models/export/exportJobHandler.ts
@@ -1,40 +1,52 @@
 import path from 'path';
+import { Feature, MultiPolygon, Polygon } from 'geojson';
+import { ogr2ogr } from 'ogr2ogr';
 import { inject, injectable } from 'tsyringe';
-import { CallbackExportResponse, callbackExportResponseSchema, CallbacksStatus, CleanupData, RasterLayerMetadata } from '@map-colonies/raster-shared';
+import {
+  CallbackExportResponse,
+  callbackExportResponseSchema,
+  CallbacksStatus,
+  CleanupData,
+  ExportFinalizeFullProcessingParams,
+  ExportFinalizeType,
+  RasterLayerMetadata,
+} from '@map-colonies/raster-shared';
 import { type Logger } from '@map-colonies/js-logger';
 import { context, trace, Tracer } from '@opentelemetry/api';
 import { ArtifactRasterType } from '@map-colonies/types';
 import { OperationStatus, TaskHandler as QueueClient } from '@map-colonies/mc-priority-queue';
 import {
-  CompletedOrFailedStatus,
   EXPORT_FAILURE_MESSAGE,
   EXPORT_SUCCESS_MESSAGE,
   GPKG_CONTENT_TYPE,
   GPKGS_PREFIX,
+  JSON_CONTENT_TYPE,
   SERVICES,
   StorageProvider,
 } from '../../../common/constants';
 import { JobHandler } from '../jobHandler';
 import { TaskMetrics } from '../../../utils/metrics/taskMetrics';
-import { ExportTask, ExportTaskParameters, IConfig, IJobHandler, JobAndTaskTelemetry } from '../../../common/interfaces';
 import {
-  ExportJob,
-  ExportInitTask,
-  ExportFinalizeTask,
-  ExportFinalizeTaskParams,
-  ExportFinalizeSuccessTaskParams,
-  ExportFinalizeFailureTaskParams,
-} from '../../../utils/zod/schemas/job.schema';
+  AggregationLayerMetadata,
+  ExportFinalizeExecutionContext,
+  ExportTask,
+  ExportTaskParameters,
+  GpkgArtifactProperties,
+  IConfig,
+  IJobHandler,
+  JsonArtifactProperties,
+} from '../../../common/interfaces';
+import { ExportJob, ExportInitTask, ExportFinalizeTask, ExportFinalizeTaskParams, exportJobSchema } from '../../../utils/zod/schemas/job.schema';
 import { S3Service } from '../../../utils/storage/s3Service';
 import { CatalogClient } from '../../../httpClients/catalogClient';
 import { internalIdSchema } from '../../../utils/zod/schemas/jobParameters.schema';
 import { ExportTaskManager } from '../../../task/models/exportTaskManager';
-import { GeoPackageClient } from '../../../utils/db/geoPackageClient';
 import { FSService } from '../../../utils/storage/fsService';
 import { createExpirationDate } from '../../../utils/dateUtil';
 import { CallbackClient } from '../../../httpClients/callbackClient';
 import { JobTrackerClient } from '../../../httpClients/jobTrackerClient';
 import { PolygonPartsMangerClient } from '../../../httpClients/polygonPartsMangerClient';
+import { convertObjectKeysToSnakeCase } from '../../../utils/db/dbUtils';
 
 @injectable()
 export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJob, ExportInitTask, ExportJob, ExportFinalizeTask> {
@@ -54,11 +66,10 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
     private readonly s3Service: S3Service,
     private readonly fsService: FSService,
     private readonly callbackClient: CallbackClient,
-    private readonly gpkgService: GeoPackageClient,
     private readonly taskMetrics: TaskMetrics,
     private readonly polygonPartsManagerClient: PolygonPartsMangerClient
   ) {
-    super(logger, queueClient, jobTrackerClient);
+    super(logger, config, queueClient, jobTrackerClient);
     this.exportTaskType = config.get<string>('jobManagement.export.tasks.tilesExporting.type');
     this.gpkgsPath = config.get<string>('jobManagement.polling.jobs.export.gpkgsPath');
     // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -111,92 +122,125 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
     await context.with(
       trace.setSpan(context.active(), this.tracer.startSpan(`${ExportJobHandler.name}.${this.handleJobFinalize.name}`)),
       async () => {
-        const activeSpan = trace.getActiveSpan();
-        const monitorAttributes = { jobId: job.id, taskId: task.id, jobType: job.type, taskType: task.type };
-        const logger = this.logger.child(monitorAttributes);
-        const taskProcessTracking = this.taskMetrics.trackTaskProcessing(job.type, task.type);
+        const context = this.createFinalizeExecutionContext(job, task);
+        const { logger, telemetry, paths } = context;
+        const { tracingSpan } = telemetry;
 
+        let finalizeParams: ExportFinalizeTaskParams = task.parameters;
         try {
-          activeSpan?.setAttributes(monitorAttributes);
+          logger.info({ msg: `Handling ${job.type} job finalization`, taskType: task.type, finalizeParams });
 
-          const gpkgRelativePath = job.parameters.additionalParams.packageRelativePath;
-          const gpkgFilePath = path.join(this.gpkgsPath, gpkgRelativePath);
-          const gpkgDirPath = path.dirname(gpkgFilePath);
-          let finalizeParams = task.parameters;
+          switch (finalizeParams.type) {
+            case ExportFinalizeType.Error_Callback:
+              logger.warn({ msg: `Finalize task type is ${task.parameters.type}, skipping finalize steps and sending failure callbacks` });
+              break;
 
-          logger.info({ msg: `Handling ${job.type} job finalization`, taskType: task.type });
-
-          if (finalizeParams.status === OperationStatus.FAILED) {
-            await this.handleFailedFinalizeTask(job, task, finalizeParams, gpkgDirPath, {
-              taskTracker: taskProcessTracking,
-              tracingSpan: activeSpan,
-            });
-            return;
-          }
-
-          if (!finalizeParams.gpkgModified) {
-            activeSpan?.addEvent('gpkg.modification.started');
-            finalizeParams = await this.modifyGpkgFile(gpkgFilePath, job, task.id, finalizeParams);
-            activeSpan?.addEvent('gpkg.modification.completed', { success: finalizeParams.gpkgModified });
-          }
-
-          const shouldUploadToS3 = this.isS3GpkgProvider && finalizeParams.gpkgModified && !finalizeParams.gpkgUploadedToS3;
-
-          logger.info({
-            msg: 'Should upload to S3',
-            shouldUploadToS3,
-            isS3GpkgProvider: this.isS3GpkgProvider,
-            gpkgModified: finalizeParams.gpkgModified,
-            gpkgUploadedToS3: finalizeParams.gpkgUploadedToS3,
-          });
-
-          if (shouldUploadToS3) {
-            activeSpan?.addEvent('s3.upload.started');
-            finalizeParams = await this.uploadGpkgToS3(gpkgFilePath, gpkgRelativePath, job.id, task.id, finalizeParams);
-            activeSpan?.addEvent('s3.upload.completed', { success: finalizeParams.gpkgUploadedToS3 });
-          }
-
-          const bypassS3OrUploaded = !this.isS3GpkgProvider || finalizeParams.gpkgUploadedToS3;
-          const gpkgProcessingComplete = finalizeParams.gpkgModified && bypassS3OrUploaded;
-
-          logger.info({
-            msg: 'GPKG processing completed',
-            success: gpkgProcessingComplete,
-            gpkgModified: finalizeParams.gpkgModified,
-            bypassS3OrUploaded,
-          });
-
-          if (gpkgProcessingComplete && !finalizeParams.callbacksSent) {
-            activeSpan?.addEvent('callbacks.sending.started');
-            await this.sendCallbacks(job, task.id, finalizeParams, gpkgDirPath);
-            activeSpan?.addEvent('callbacks.sending.completed');
-          }
-
-          if (gpkgProcessingComplete) {
-            activeSpan?.addEvent('all.steps.completed');
-            logger.info({ msg: 'All finalize steps completed successfully' });
-            await this.completeTask(job, task, { taskTracker: taskProcessTracking, tracingSpan: activeSpan });
+            case ExportFinalizeType.Full_Processing:
+              finalizeParams = await this.handleFullProcessing(context, finalizeParams);
+              break;
           }
         } catch (err) {
-          await this.handleError(err, job, task, { taskTracker: taskProcessTracking, tracingSpan: activeSpan });
+          await this.handleError(err, job, task, telemetry);
         } finally {
-          activeSpan?.end();
+          const updatedJob = await this.queueClient.jobManagerClient.getJob(job.id);
+          const validJob = exportJobSchema.parse(updatedJob);
+          const isTerminalStatus = validJob.status === OperationStatus.COMPLETED || validJob.status === OperationStatus.FAILED;
+          const isErrorCallback = finalizeParams.type === ExportFinalizeType.Error_Callback;
+          const shouldSendCallbacks = (isTerminalStatus || isErrorCallback) && !finalizeParams.callbacksSent;
+          logger.info({ msg: 'should send callbacks', shouldSendCallbacks, isTerminalStatus, isErrorCallback });
+          if (shouldSendCallbacks) {
+            tracingSpan?.addEvent('callbacks.sending.started');
+            await this.sendCallbacks(validJob, task.id, finalizeParams, paths.gpkgDirPath);
+            tracingSpan?.addEvent('callbacks.sending.completed');
+            if (isErrorCallback) {
+              logger.info({ msg: 'Finalize task type is Error_Callback, Completing after callbacks sends' });
+              await this.completeTask(job, task, telemetry);
+            }
+          }
+          tracingSpan?.end();
         }
       }
     );
   }
 
-  private async handleFailedFinalizeTask(
-    job: ExportJob,
-    task: ExportFinalizeTask,
-    finalizeParams: ExportFinalizeFailureTaskParams,
-    gpkgDirPath: string,
-    telemetry: JobAndTaskTelemetry
-  ): Promise<void> {
-    this.logger.info({ msg: 'Processing failed finalize task' });
-    await this.updateCallbackParams(job, OperationStatus.FAILED, { errorReason: EXPORT_FAILURE_MESSAGE });
-    await this.sendCallbacks(job, task.id, finalizeParams, gpkgDirPath);
-    await this.completeTask(job, task, telemetry);
+  private createFinalizeExecutionContext(job: ExportJob, task: ExportFinalizeTask): ExportFinalizeExecutionContext {
+    const activeSpan = trace.getActiveSpan();
+    const monitorAttributes = { jobId: job.id, taskId: task.id, jobType: job.type, taskType: task.type };
+    const taskTracker = this.taskMetrics.trackTaskProcessing(job.type, task.type);
+    const logger = this.logger.child(monitorAttributes);
+
+    activeSpan?.setAttributes(monitorAttributes);
+
+    const gpkgRelativePath = job.parameters.additionalParams.packageRelativePath;
+    const gpkgFilePath = path.join(this.gpkgsPath, gpkgRelativePath);
+    const gpkgDirPath = path.dirname(gpkgFilePath);
+
+    return {
+      job,
+      task,
+      paths: {
+        gpkgFilePath,
+        gpkgRelativePath,
+        gpkgDirPath,
+      },
+      telemetry: {
+        taskTracker,
+        tracingSpan: activeSpan,
+      },
+      logger,
+    };
+  }
+
+  private async handleFullProcessing(
+    context: ExportFinalizeExecutionContext,
+    taskParams: ExportFinalizeFullProcessingParams
+  ): Promise<ExportFinalizeTaskParams> {
+    const { job, task, paths, telemetry, logger } = context;
+    const { taskTracker, tracingSpan: activeSpan } = telemetry;
+    const { gpkgFilePath, gpkgRelativePath } = paths;
+
+    let fullProcessingParams = taskParams;
+    let gpkgProcessingComplete = false;
+
+    if (!fullProcessingParams.gpkgModified) {
+      activeSpan?.addEvent('gpkg.modification.started');
+      fullProcessingParams = await this.modifyGpkgFile(gpkgFilePath, job, task.id, fullProcessingParams);
+      activeSpan?.addEvent('gpkg.modification.completed', { success: fullProcessingParams.gpkgModified });
+    }
+
+    const shouldUploadToS3 = this.isS3GpkgProvider && fullProcessingParams.gpkgModified && !fullProcessingParams.gpkgUploadedToS3;
+
+    logger.info({
+      msg: 'Should upload to S3',
+      shouldUploadToS3,
+      isS3GpkgProvider: this.isS3GpkgProvider,
+      gpkgModified: fullProcessingParams.gpkgModified,
+      gpkgUploadedToS3: fullProcessingParams.gpkgUploadedToS3,
+    });
+
+    if (shouldUploadToS3) {
+      activeSpan?.addEvent('s3.upload.started');
+      fullProcessingParams = await this.uploadGpkgToS3(gpkgFilePath, gpkgRelativePath, job.id, task.id, fullProcessingParams);
+      activeSpan?.addEvent('s3.upload.completed', { success: fullProcessingParams.gpkgUploadedToS3 });
+    }
+
+    const bypassS3OrUploaded = !this.isS3GpkgProvider || fullProcessingParams.gpkgUploadedToS3;
+    gpkgProcessingComplete = fullProcessingParams.gpkgModified && bypassS3OrUploaded;
+
+    logger.info({
+      msg: 'GPKG processing completed',
+      success: gpkgProcessingComplete,
+      gpkgModified: fullProcessingParams.gpkgModified,
+      bypassS3OrUploaded,
+    });
+
+    if (gpkgProcessingComplete) {
+      activeSpan?.addEvent('all.steps.completed');
+      logger.info({ msg: 'All finalize steps completed successfully' });
+      await this.completeTask(job, task, { taskTracker, tracingSpan: activeSpan });
+    }
+
+    return fullProcessingParams;
   }
 
   private createExportTask(job: ExportJob, metadata: RasterLayerMetadata): ExportTask {
@@ -234,40 +278,102 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
     gpkgFilePath: string,
     job: ExportJob,
     taskId: string,
-    taskParams: ExportFinalizeSuccessTaskParams
-  ): Promise<ExportFinalizeSuccessTaskParams> {
-    this.logger.info({ msg: 'Modify gpkg file (create table from metadata)', jobId: job.id, taskId, gpkgFilePath });
+    taskParams: ExportFinalizeFullProcessingParams
+  ): Promise<ExportFinalizeFullProcessingParams> {
+    const logger = this.logger.child({ jobId: job.id, taskId });
 
+    const validInternalId = internalIdSchema.parse(job).internalId;
+    logger.debug({ msg: 'internalId validation passed', internalId: validInternalId });
+
+    const layerName = job.parameters.additionalParams.polygonPartsEntityName;
     const featureCollectionFilter = job.parameters.exportInputParams.roi;
     const aggregatedMetadata = await this.polygonPartsManagerClient.getAggregatedLayerMetadata(
       job.parameters.additionalParams.polygonPartsEntityName,
       featureCollectionFilter
     );
-    this.logger.info({ msg: 'Received Aggregated layer metadata', jobId: job.id, taskId, aggregatedMetadata });
+    logger.info({ msg: 'Received Aggregated layer metadata', aggregatedMetadata });
+    const { metadata } = await this.catalogClient.findLayer(validInternalId);
 
-    const isTableCreated = this.gpkgService.createTableFromMetadata(gpkgFilePath, aggregatedMetadata);
-    if (isTableCreated) {
-      const fileSize = await this.fsService.getFileSize(gpkgFilePath);
-      await this.updateCallbackParams(job, OperationStatus.IN_PROGRESS, { fileSize });
-      const updatedParams = await this.markFinalizeStepAsCompleted(job.id, taskId, taskParams, 'gpkgModified');
-      return updatedParams;
-    }
-    return taskParams;
+    const roiMetadata = this.createRoiMetadata(metadata, aggregatedMetadata);
+    const gpkgMetadata = this.createGpkgMetadata(roiMetadata);
+
+    //TODO In future, we will remove the json metadata file and support only gpkg
+    const jsonPath = gpkgFilePath.replace(/\.gpkg$/, '.json');
+    logger.info({ msg: 'Creating json metadata file', jsonPath });
+
+    logger.info({ msg: 'Modify gpkg file (create table from metadata)', gpkgFilePath, gpkgMetadata });
+
+    await ogr2ogr(
+      { ...gpkgMetadata },
+      {
+        format: 'GPKG',
+        destination: gpkgFilePath,
+        options: ['-nln', `${layerName}_metadata`, '-overwrite'],
+      }
+    );
+
+    const gpkgSha256 = await this.fsService.calculateFileSha256(gpkgFilePath);
+    const jsonMetadata = this.prepareJsonFileMetadata(roiMetadata, gpkgSha256);
+    await this.fsService.uploadJsonFile(jsonPath, jsonMetadata);
+
+    const fileSize = await this.fsService.getFileSize(gpkgFilePath);
+    const jsonFileSize = await this.fsService.getFileSize(jsonPath); //TODO: In future, we will remove the json metadata file and support only gpkg
+
+    logger.info({ msg: 'Gpkg file size', fileSize });
+    logger.info({ msg: 'Json file size', jsonFileSize });
+    await this.updateCallbackParams(job, OperationStatus.IN_PROGRESS, { fileSize, jsonFileData: { sha256: gpkgSha256, size: jsonFileSize } });
+    const updatedParams = await this.markFinalizeStepAsCompleted(job.id, taskId, taskParams, 'gpkgModified');
+    return updatedParams;
+  }
+
+  private createRoiMetadata(
+    layerMetadata: RasterLayerMetadata,
+    aggregatedMetadata: AggregationLayerMetadata
+  ): Feature<Polygon | MultiPolygon, GpkgArtifactProperties> {
+    const { footprint: layerFootprint, productStatus, ...relevantMetadata } = layerMetadata;
+    const { footprint: aggregatedFootprint, ...relevantAggregatedMetadata } = aggregatedMetadata;
+    return {
+      type: 'Feature',
+      properties: {
+        ...relevantMetadata,
+        ...relevantAggregatedMetadata,
+      },
+      geometry: aggregatedFootprint,
+    };
+  }
+
+  private prepareJsonFileMetadata(feature: Feature<Polygon | MultiPolygon, GpkgArtifactProperties>, gpkgSha256: string): JsonArtifactProperties {
+    return {
+      ...feature.properties,
+      footprint: feature.geometry,
+      sha256: gpkgSha256,
+    };
+  }
+
+  private createGpkgMetadata(feature: Feature<Polygon | MultiPolygon, GpkgArtifactProperties>): Feature<Polygon | MultiPolygon> {
+    const snakeCaseProps = convertObjectKeysToSnakeCase(feature.properties);
+    return {
+      ...feature,
+      properties: snakeCaseProps,
+    };
   }
 
   private async uploadGpkgToS3(
-    gpkgFilePath: string,
+    gpkgPath: string,
     gpkgRelativePath: string,
     jobId: string,
     taskId: string,
-    taskParams: ExportFinalizeSuccessTaskParams
-  ): Promise<ExportFinalizeSuccessTaskParams> {
-    const logger = this.logger.child({ jobId, taskId });
+    taskParams: ExportFinalizeFullProcessingParams
+  ): Promise<ExportFinalizeFullProcessingParams> {
+    const gpkgS3Key = `${GPKGS_PREFIX}/${gpkgRelativePath}`;
+    const jsonS3Key = gpkgS3Key.replace(/\.gpkg$/, '.json'); //TODO: In future, we will remove the json metadata file and support only gpkg
+    const jsonPath = gpkgPath.replace(/\.gpkg$/, '.json'); //TODO: In future, we will remove the json metadata file and support only gpkg
 
-    const s3Key = `${GPKGS_PREFIX}/${gpkgRelativePath}`;
-    logger.info({ msg: 'Upload gpkg file to S3', gpkgFilePath, s3Key, contentType: GPKG_CONTENT_TYPE });
-    await this.s3Service.uploadFile(gpkgFilePath, s3Key, GPKG_CONTENT_TYPE);
-    await this.fsService.deleteFileAndParentDir(gpkgFilePath);
+    await this.s3Service.uploadFiles([
+      { filePath: gpkgPath, s3Key: gpkgS3Key, contentType: GPKG_CONTENT_TYPE },
+      { filePath: jsonPath, s3Key: jsonS3Key, contentType: JSON_CONTENT_TYPE },
+    ]);
+    await this.fsService.deleteFileAndParentDir(gpkgPath);
     const updatedParams = await this.markFinalizeStepAsCompleted(jobId, taskId, taskParams, 'gpkgUploadedToS3');
     return updatedParams;
   }
@@ -284,11 +390,11 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
     await this.queueClient.jobManagerClient.updateJob(job.id, { parameters: { ...job.parameters } });
   }
 
-  private async sendCallbacks(job: ExportJob, taskId: string, taskParams: ExportFinalizeTaskParams, dirPath: string): Promise<boolean> {
+  private async sendCallbacks(job: ExportJob, taskId: string, taskParams: ExportFinalizeTaskParams, dirPath: string): Promise<void> {
     try {
       const cleanupExpirationTimeUTC = createExpirationDate(this.cleanupExpirationDays);
       const cleanupDataParams: CleanupData = { cleanupExpirationTimeUTC, directoryPath: dirPath };
-      const callbackParams = this.createCallbacksParams(job, taskParams.status, cleanupExpirationTimeUTC);
+      const callbackParams = this.createCallbacksParams(job, cleanupExpirationTimeUTC);
 
       await this.queueClient.jobManagerClient.updateJob(job.id, {
         parameters: { ...job.parameters, cleanupDataParams, callbackParams },
@@ -297,11 +403,11 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
       const targetCallbacks = job.parameters.exportInputParams.callbackUrls;
       // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
       if (!targetCallbacks?.length) {
-        this.logger.info({ msg: 'No callbacks url provided, skipping callbacks sending', jobId: job.id });
-        return false;
+        this.logger.warn({ msg: 'No callbacks url provided, skipping callbacks sending', jobId: job.id });
+        return;
       }
 
-      this.logger.info({ msg: 'Sending callbacks', jobId: job.id, status: taskParams.status, callbackCount: targetCallbacks.length });
+      this.logger.info({ msg: 'Sending callbacks', jobId: job.id, callbackCount: targetCallbacks.length });
 
       await Promise.all(
         targetCallbacks.map(async (callback) =>
@@ -317,31 +423,32 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
         )
       );
 
-      const updatedParams = await this.markFinalizeStepAsCompleted(job.id, taskId, taskParams, 'callbacksSent');
-      return updatedParams.callbacksSent;
+      await this.markFinalizeStepAsCompleted(job.id, taskId, taskParams, 'callbacksSent');
     } catch (err) {
       this.logger.error({ msg: 'Sending callbacks has failed', err });
-      throw err;
     }
   }
 
-  private createCallbacksParams(job: ExportJob, status: CompletedOrFailedStatus, expirationDate: Date): CallbackExportResponse {
+  private createCallbacksParams(job: ExportJob, expirationDate: Date): CallbackExportResponse {
     const { additionalParams, callbackParams } = job.parameters;
     const { packageRelativePath, fileNamesTemplates } = additionalParams;
     const validCallbackParams = callbackExportResponseSchema.parse(callbackParams);
+    const { jsonFileData, ...clientsCallbackParams } = validCallbackParams;
 
     const callbackResponse: CallbackExportResponse = {
-      ...validCallbackParams,
-      status,
+      ...clientsCallbackParams,
+      status: OperationStatus.FAILED,
       expirationTime: expirationDate,
       artifacts: [],
       links: undefined,
+      description: EXPORT_FAILURE_MESSAGE,
     };
 
-    if (status === OperationStatus.COMPLETED) {
+    if (job.status === OperationStatus.COMPLETED) {
       const gpkgDownloadUrl = this.isS3GpkgProvider
         ? `${this.downloadUrl}/${GPKGS_PREFIX}/${packageRelativePath}`
-        : `${this.downloadUrl}/${packageRelativePath}`; // later when we change download server mount directory, the path for s3 and fs should be the same
+        : `${this.downloadUrl}/${packageRelativePath}`; //TODO: later when we change download server mount directory, the path for s3 and fs should be the same
+      const jsonDownloadUrl = gpkgDownloadUrl.replace(/\.gpkg$/, '.json'); //TODO: In future, we will remove the json metadata file and support only gpkg
 
       callbackResponse.artifacts = [
         {
@@ -349,11 +456,20 @@ export class ExportJobHandler extends JobHandler implements IJobHandler<ExportJo
           type: ArtifactRasterType.GPKG,
           size: validCallbackParams.fileSize ?? 0,
           url: gpkgDownloadUrl,
+          sha256: validCallbackParams.jsonFileData?.sha256,
+        },
+        {
+          //TODO: In future, we will remove the json metadata file and support only gpkg
+          name: fileNamesTemplates.packageName,
+          type: ArtifactRasterType.METADATA,
+          size: validCallbackParams.jsonFileData?.size ?? 0,
+          url: jsonDownloadUrl,
         },
       ];
 
-      callbackResponse.links = { dataURI: gpkgDownloadUrl };
+      callbackResponse.status = OperationStatus.COMPLETED;
       callbackResponse.description = EXPORT_SUCCESS_MESSAGE;
+      callbackResponse.links = { dataURI: gpkgDownloadUrl, metadataURI: jsonDownloadUrl };
     }
 
     return callbackResponse;