+++ src/utils/storage/fsService.ts
@@ -1,4 +1,7 @@
 import fs from 'fs/promises';
+import { createReadStream } from 'fs';
+import { finished } from 'stream/promises'; // Promise-based stream completion
+import crypto from 'crypto';
 import path from 'path';
 import { inject, injectable } from 'tsyringe';
 import { Logger } from '@map-colonies/js-logger';
@@ -9,6 +12,39 @@ import { FSError } from '../../common/errors';
 @injectable()
 export class FSService {
   public constructor(@inject(SERVICES.LOGGER) private readonly logger: Logger, @inject(SERVICES.TRACER) private readonly tracer: Tracer) {}
+
+  public async uploadJsonFile(filePath: string, data: Record<string, unknown>): Promise<void> {
+    return context.with(trace.setSpan(context.active(), this.tracer.startSpan(`${FSService.name}.${this.uploadJsonFile.name}`)), async () => {
+      const activeSpan = trace.getActiveSpan();
+      try {
+        activeSpan?.setAttributes({
+          filePath,
+          operation: 'uploadJsonFile',
+        });
+
+        this.logger.info({ msg: 'Uploading JSON file', filePath });
+
+        // eslint-disable-next-line @typescript-eslint/no-magic-numbers
+        await fs.writeFile(filePath, JSON.stringify(data, null, 2));
+
+        activeSpan?.addEvent('file.uploaded', { filePath });
+        this.logger.info({ msg: 'JSON file uploaded successfully', filePath });
+      } catch (err) {
+        const error = new FSError(err, `Failed to upload JSON file ${filePath}`);
+
+        this.logger.error({ msg: error.message, error });
+        activeSpan?.recordException(error);
+        activeSpan?.setStatus({
+          code: SpanStatusCode.ERROR,
+          message: error.message,
+        });
+
+        throw error;
+      } finally {
+        activeSpan?.end();
+      }
+    });
+  }
 
   public async deleteFile(filePath: string): Promise<void> {
     return context.with(trace.setSpan(context.active(), this.tracer.startSpan(`${FSService.name}.${this.deleteFile.name}`)), async () => {
@@ -149,6 +185,47 @@ export class FSService {
           message: error.message,
         });
 
+        throw error;
+      } finally {
+        activeSpan?.end();
+      }
+    });
+  }
+
+  public async calculateFileSha256(filePath: string): Promise<string> {
+    return context.with(trace.setSpan(context.active(), this.tracer.startSpan(`${FSService.name}.${this.calculateFileSha256.name}`)), async () => {
+      const activeSpan = trace.getActiveSpan();
+      try {
+        activeSpan?.setAttributes({
+          filePath,
+          operation: 'calculateFileSha256',
+        });
+
+        this.logger.info({ msg: 'Calculating file SHA256', filePath });
+        await fs.access(filePath);
+
+        const hash = crypto.createHash('sha256');
+        const fileStream = createReadStream(filePath);
+
+        fileStream.on('data', (data) => {
+          hash.update(data);
+        });
+
+        await finished(fileStream);
+
+        const sha256 = hash.digest('hex');
+
+        activeSpan?.addEvent('file.sha256.calculated', { sha256 });
+        this.logger.info({ msg: 'SHA256 calculated successfully', filePath, sha256 });
+        return sha256;
+      } catch (err) {
+        const error = new FSError(err, `Failed to calculate SHA256 for ${filePath}`);
+        this.logger.error({ msg: error.message, error });
+        activeSpan?.recordException(error);
+        activeSpan?.setStatus({
+          code: SpanStatusCode.ERROR,
+          message: error.message,
+        });
         throw error;
       } finally {
         activeSpan?.end();